packages/aop/pyproject.toml
[project]
name = "cogency-aop"
version = "0.0.1"
description = "AOP (Agent Operating Procedure) DSL parser, compiler, executor"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.9",
    "PyYAML>=6.0",
    "cogency-schemas",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aop"]
